make task list recursive
